Split ui_utils.py, reduce logging
